Modular decom implemented. Needs tests
