+++ openc3/lib/openc3.rb
@@ -37,6 +37,7 @@ require 'openc3/version'
 require 'openc3/top_level'
 require 'openc3/core_ext'
 require 'openc3/utilities'
+require 'openc3/accessors'
 require 'openc3/conversions'
 require 'openc3/interfaces'
 require 'openc3/processors'
