+++ openc3/lib/openc3/accessors/html_accessor.rb
@@ -1,0 +1,28 @@
+# encoding: ascii-8bit
+
+# Copyright 2022 OpenC3, Inc.
+# All Rights Reserved.
+#
+# This program is free software; you can modify and/or redistribute it
+# under the terms of the GNU Affero General Public License
+# as published by the Free Software Foundation; version 3 with
+# attribution addendums as found in the LICENSE.txt
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+
+require 'openc3/accessors/xml_accessor'
+
+module OpenC3
+  class HtmlAccessor < XmlAccessor
+    def self.buffer_to_doc(buffer)
+      Nokogiri.HTML(buffer)
+    end
+
+    def self.doc_to_buffer(doc)
+      doc.to_html
+    end
+  end
+end
