+++ openc3/lib/openc3/models/cvt_model.rb
@@ -26,16 +26,7 @@ module OpenC3
     @overrides = {}
 
     def self.build_json_from_packet(packet)
-      json_hash = {}
-      packet.sorted_items.each do |item|
-        json_hash[item.name] = packet.read_item(item, :RAW)
-        json_hash["#{item.name}__C"] = packet.read_item(item, :CONVERTED) if item.read_conversion or item.states
-        json_hash["#{item.name}__F"] = packet.read_item(item, :FORMATTED) if item.format_string
-        json_hash["#{item.name}__U"] = packet.read_item(item, :WITH_UNITS) if item.units
-        limits_state = item.limits.state
-        json_hash["#{item.name}__L"] = limits_state if limits_state
-      end
-      json_hash
+      packet.decom
     end
 
     # Delete the current value table for a target
