+++ openc3/spec/packets/structure_item_spec.rb
@@ -221,9 +221,10 @@ module OpenC3
       it "creates a Hash" do
         item = StructureItem.new("test", 0, 8, :UINT, :BIG_ENDIAN, 16)
         hash = item.as_json(:allow_nan => true)
-        expect(hash.keys.length).to eql 7
+        expect(hash.keys.length).to eql 8
         expect(hash.keys).to include('name', 'bit_offset', 'bit_size', 'data_type', 'endianness', 'array_size', 'overflow')
         expect(hash["name"]).to eql "TEST"
+        expect(hash["key"]).to eql "test"
         expect(hash["bit_offset"]).to eql 0
         expect(hash["bit_size"]).to eql 8
         expect(hash["data_type"]).to eql :UINT
